+++ CardPayments/src/main/java/com/paypal/android/cardpayments/api/CreateOrderResponse.kt
@@ -1,5 +1,0 @@
-package com.paypal.android.cardpayments.api
-
-import com.paypal.android.corepayments.OrderStatus
-
-internal data class CreateOrderResponse(val orderId: String, val status: OrderStatus)