+++ CHANGELOG.md
@@ -5,6 +5,10 @@
   * Bump Kotlin version to `1.8.21`
 * CardPayments
   * Make `Card` implement `Parcelable`
+  * Make `CardRequest` implement `Parcelable`
+  * Add the following properties to `CardRequest`:
+    * `shouldVault`
+    * `vaultCusomterId`
 * CorePayments
   * Make `Address` implement `Parcelable`
 * PayPalNativePayments